+++ trade-service/src/main/java/com/moni/trade/asset/application/service/AssetService.java
@@ -209,7 +209,7 @@ public class AssetService {
 
         return new HoldingInput(
                 holding.ticker(),
-                holding.quantity(),
+                holding.quantity().longValue(),
                 holding.averagePrice(),
                 holding.totalAmount()
         );
